cleanup(bigquery): centralize test helpers
